Return error if user-supplied task name already exists
Per the docs: https://cloud.google.com/tasks/docs/reference/rest/v2/projects.locations.queues.tasks/create#body.request_body.FIELDS.task

> Explicitly specifying a task ID enables task de-duplication. If
a task's ID is identical to that of an existing task or a task that
was deleted or executed recently then the call will fail with
`google.rpc.Code.ALREADY_EXISTS`. If the task's queue was created
using Cloud Tasks, then another task with the same name can't be
created for ~1hour after the original task was deleted or executed.
If the task's queue was created using queue.yaml or queue.xml,
then another task with the same name can't be created for ~9days
after the original task was deleted or executed.

For the purposes of the emulator, I don't think it's necessary to
model the time-based garbage collection of old task names. A task
name that has been used since the emulator was last booted / purged
will be rejected.